Aerobot/requirements.txt
streamlit
pandas
numpy
plotly
firebase-admin
google-generativeai
torch
requests
toml